+++ aws_lambda_powertools/utilities/kafka/serialization/base.py
@@ -4,6 +4,8 @@ from abc import ABC, abstractmethod
 from typing import TYPE_CHECKING, Any
 
 if TYPE_CHECKING:
+    from collections.abc import Callable
+
     from aws_lambda_powertools.utilities.kafka.serialization.types import T
 
 
@@ -16,21 +18,21 @@ class OutputSerializerBase(ABC):
 
     Methods
     -------
-    serialize(data, output_class)
+    serialize(data, output)
         Abstract method that must be implemented by subclasses to serialize data.
 
     Examples
     --------
     >>> class MyOutputSerializer(OutputSerializerBase):
-    ...     def serialize(self, data: dict[str, Any], output_class=None):
-    ...         if output_class:
+    ...     def serialize(self, data: dict[str, Any], output=None):
+    ...         if output:
     ...             # Convert dictionary to class instance
-    ...             return output_class(**data)
+    ...             return output(**data)
     ...         return data  # Return as is if no output class provided
     """
 
     @abstractmethod
-    def serialize(self, data: dict[str, Any], output_class: type[T] | None = None) -> T | dict[str, Any]:
+    def serialize(self, data: dict[str, Any], output: type[T] | Callable | None = None) -> T | dict[str, Any]:
         """
         Serialize dictionary data into a specific output format or class instance.
 
@@ -41,14 +43,14 @@ class OutputSerializerBase(ABC):
         ----------
         data : dict[str, Any]
             The dictionary data to serialize.
-        output_class : type[T] or None, optional
+        output : type[T] or None, optional
             Optional class type to convert the dictionary into. If provided,
             the method should return an instance of this class.
 
         Returns
         -------
         T or dict[str, Any]
-            An instance of output_class if provided, otherwise a processed dictionary.
-            The generic type T represents the type of the output_class.
+            An instance of output if provided, otherwise a processed dictionary.
+            The generic type T represents the type of the output.
         """
         raise NotImplementedError("Subclasses must implement this method")
