+++ tests/functional/kafka_consumer/_protobuf/test_kafka_consumer_with_protobuf.py
@@ -217,11 +217,8 @@ def test_kafka_consumer_with_custom_object(
     """Test Kafka consumer with Protobuf deserialization and custom object serialization."""
 
     # Define a custom output object class
-    class UserCustomObject:
-        def __init__(self, proto_message):
-            self.name = proto_message.name
-            self.age = proto_message.age
-            self.custom_field = f"{proto_message.name} is {proto_message.age} years old"
+    def dict_output(data: dict) -> dict:
+        return data
 
     # Create dict to capture results
     result_data = {}
@@ -229,17 +226,15 @@ def test_kafka_consumer_with_custom_object(
     schema_config = SchemaConfig(
         value_schema_type="PROTOBUF",
         value_schema=User,
-        value_output_serializer=lambda msg: UserCustomObject(msg),
+        value_output_serializer=dict_output,
     )
 
     @kafka_consumer(schema_config=schema_config)
     def handler(event: ConsumerRecords, context):
         # Capture the results to verify
         record = next(event.records)
-        result_data["value_type"] = type(record.value).__name__
-        result_data["name"] = record.value.name
-        result_data["age"] = record.value.age
-        result_data["custom_field"] = record.value.custom_field
+        result_data["name"] = record.value.get("name")
+        result_data["age"] = record.value.get("age")
         return {"processed": True}
 
     # Call the handler
@@ -247,10 +242,8 @@ def test_kafka_consumer_with_custom_object(
 
     # Verify the results
     assert result == {"processed": True}
-    assert result_data["value_type"] == "UserCustomObject"
     assert result_data["name"] == "John Doe"
     assert result_data["age"] == 30
-    assert result_data["custom_field"] == "John Doe is 30 years old"
 
 
 def test_kafka_consumer_with_multiple_records(lambda_context):
